Skeleton of an ARM assembly language file.
The template code that Chadwick has written will start at
the code section labeled ".init", so this tells the program
that our Operating System program should start here.

The following two lines ".globl _start" and "_start:" are
there to surpress a warning. All of the code so far are
instructions for the assembler and don't correspond to
instructions to the Raspberry Pi.