Prepare the LED GPIO pin to be turned on.
"mov r1,#1" is faster than "ldr r1,=0x00000001", and is
used to store the value 1 in the Register 1.

"lsl r1,#18" takes the value stored in Register 1 and
shifts it to the left by 18 places.

"str r1,[r0,#4]" stores the value in Register 1 into
the memory address calculated from the sum of the value
in Register 0 (0x20200000) and the value 4.