Turn on the LED GPIO pin.
Like before, "mov r1,#1" and "lsl r1,#16" place a value
of 1 in Register 1 and shift it to the left a number of
places - in this case, 16.

"str r1,[r0,#40]" stores the value in Register 1 at the
address calculated from the sum of the value in Register 0
(0x20200000) and the value 40.