Enter an infinite loop to idle the processor.
"loop$:" is a label corresponding to a block of code
and can be referred to later on in the program. The '$'
signifies to other programs that this label is not important
outside of this block of code.

"b loop$" uses the branch command and tells the processor to
jump back to the block of code labeled "loop$". This causes
the processor to infinitely loop until it is safely powered off.